StockItem.displayqty is now a Numeric() not an Int()
It's always been an inconsistency that stock on display stock lines
could only be sold in integer units - and a cause of some problems if
the stock items on the line ever had a non-integer amount of usage
recorded against them.

Change displayqty to be the same type as all other quantities.