Add column to stocktypes specifying default sale unit
Until now, the default quantity when selling stock has been "1.0".  This has
worked fine for pints, measures of spirits, bottles, and so on.  But now we're
selling wine and soft drinks in units of "ml" it's a bit silly to price things by
the ml - we run into rounding problems, for a start.

This commit adds a column to the stocktypes table specifying the default quantity
to sell when this stocktype is sold; the saleprice field now refers to this
quantity rather than "1.0".

Modifiers can still adjust the price and quantity as they see fit.