WIP: database query optimisation
Remove several round-trips to the database during common sale operations
by giving the ORM hints about what we're going to do with the objects it
loads.

This may be exposing an underlying issue with keeping detached object
instances around in the register state.  This needs looking into - eg.
register.trans possibly ought to be a transaction ID not an ORM
instance.