Fix for sales from display stocklines with negative ondisplay
Sometimes a stock item can end up with a negative amount "on display" -
for example by careless use of the "record waste" function with negative
amounts of stock.

calculate_sale() would never return a negative amount to sell from an
item, but the amount by which the item was short would still be added to
the target amount!

This commit ensures that the speculative amount to sell from an item can
never be less than zero.