Store transaction ID instead of Transaction object in register
The register would keep a detached Transaction object around between ORM
sessions, and re-add it to the session when necessary.  This caused
problems when other code loaded the Transaction from the database before
entering the register code.

This commit alters the register and payment code to store the
transaction ID instead.  The Transaction object is loaded from the
database (or from the session's identity map) when needed.