Remove use of detached instances in several places
Keeping a model instance around after the end of the database session,
and re-introducing it into another database session using td.s.add() or
td.s.merge() is a pattern than can introduce difficult-to-find bugs that
depend strongly on the order that users do things in.

It's much better not to keep detached instances around at all, but
instead keep data that can be used to fetch a fresh instance using
td.s.query(model).get().  This is fairly cheap - if the model instance
has already been fetched in the current session, calling get() won't hit
the database at all.

Two of the major culprits were ui.popupfield() and ui.listfield(), which
had APIs that required the use of detached instances.  These are removed
in this commit and replaced with ui.modelpopupfield() and
ui.modellistfield() which have different APIs and do not make any use of
detached instances.  All uses of ui.popupfield() and ui.listfield() have
been updated to use the new API.

The object hierarchy in ui.py has been altered slightly.  Class field
has been separated out into two classes: field() for things that can
receive the input focus, and child class valuefield() for things that
have a value that can be set with the set() method, read with the read()
method, and which call sethook when the value is updated.