Minor updates to session total upload sequence
We had a partial upload of daily totals at the Pembury over the weekend,
which I suspect is because the till software was running in a ssh
session and the session was closed before the upload completed.

This commit changes the order in which things happen: first the totals
are committed to the database, then a "please wait, uploading..."
message appears; the totals are uploaded, and finally the session totals
printout is done.  Members of staff expect the printout, so are unlikely
to kill the session before the printout starts.