Speed up cancelling whole transactions
Add cascades to the database to delete payments, transaction lines and
stockout entries when a transaction is deleted.

Add passive_deletes=True to the appropriate relationships in the ORM so
the ORM doesn't duplicate the database's work.

This commit requires database updates.  The checkdb command will provide
a suitable update script.