Speed up transaction loading
Now that we store text for all new transaction lines, we don't need to
access the stockout and stocktypes tables while loading transactions
into the register.  Doing so slows down the database query considerably.

This one-line change speeds up transaction loading where the transaction
line text field is filled in - now the common case.  If historical
transactions are accessed where translines.text is null, loading will be
slower - but this is very uncommon.