Add command to fill in text field in translines
Currently the text column in translines may contain null values,
indicating that the text should be calculated from department and
stockout information.

In quicktill-0.13 I intend to set the column to not null, which will
simplify some database queries in the till and tillweb.  This commit
adds a command that updates old translines with appropriate text.