Improve user / token management menu
Add a new popup for editing and assigning tokens directly, without going
through the user list and existing manage-tokens-for-this-user popup.

This should be much more useful for taking an unknown token and lending
it to an existing user - no need to figure out who it currently belongs
to first, remove it from them and assign it to the new user.