Use django reverse() for all URL references in tillweb
This restores the desirable "don't repeat yourself" property for URLs in
tillweb and removes the kludges that were used before.

Model instances are annotated with the name of the pub they refer to,
and a reference to the django reverse() function, when they are loaded
from the database.

NB new model instances don't get these annotations, even when they are
added to a session.