Fix Xero upload code for delivery and session URLs
The recent changes to tillweb mean the tillweb_url property has gone
away from models.  Update the Xero upload code not to depend on these.

This is a bit of a hack (it encodes the tillweb URL structure into the
Xero upload code) but it means we don't have to pull in all of django
core on all tills.