Add an on-screen keyboard for development and testing
runtill start --debug-kbd now pops up a window with an appropriate till
keyboard where possible.  (When no till keyboard is expected, for
example when running in stock terminal mode, nothing happens.)

Keycaps for this keyboard are based on the keycap database entries.
However, changes to these are not reflected on the keyboard until the
till is restarted.