Event loop changes
Preparation for pluggable user interfaces: pluggable event loops

Two new event loop implementations are provided, one based on the
"selectors" library module introduced in python-3.4 and the other based
on GLib.

All code that makes use of the event loop (mostly for timer callbacks,
but a couple of fd watches as well) has been updated to use the new API.

By default the "selectors" event loop is used, but the "-g" flag to the
start command will switch to the GLib-based event loop.

Incidental changes:
 - a new "-n" or "--nolisten" option to the start command that
   prevents listening for user tokens
 - a couple of obsolete extras have been removed (Bar Billiards
   checker and coffee pot timer)
 - "moneyfield" has been copied from the card module to the ui module
   (but nothing uses it yet)
 - when the --debug command line option is set, tillconfig.debug
   will be True and there is a new menu option under Manage Till
   for various debug actions (raise an exception, etc.)