Pluggable user interface backends
Move things around so that the code that deals with ncurses is in a
new module, ui_ncurses.py, and is called as required by the
backend-independent code in ui.py

Change the way that colours are handled so that it doesn't depend on
ncurses colorpairs and attributes; these are handled transparently
by ui_ncurses.py.  For now we are still limited to the eight basic
colours.

Introduce a new backend that draws directly in a Gtk window using
cairo, ui_gtk.py.  This is not ready for production use.