Add 'testmenu' subcommand
Add a subcommand for interactive checking of food menu files.

Previously this was done with an entry in the configuration file, but
that was a bit kludgy because the user invoking the menu test code would
require permission to access the database.

The "old way" of starting the checker still works.