Speed up scrolling down in scrollables
The code in scrollables assumed that drawing to the screen was cheap,
and so redrew multiple times while scrolling down.

With the gtk backend, drawing is less cheap so suppress it while working
out how to scroll down to get the cursor back on screen.  Also cache
wrapped text so we don't have to re-run the linewrap routine multiple
times on the same text, and suppress drawing of strings that are just
spaces.