Exit with "idle" exit code on update notification
When the till is provided with an exit code to use for "exit on idle",
add a database notification listener for "update" notifications, and
exit with the idle exit code when one is received.

In usual deployments this will lead to the till software being restarted
to pick up a new version, config file update, etc.

Closes #177 on github.