Allow department to be changed while editing stock type
Changing the department of a stocktype used to be forbidden because it
could make stock on sale be inconsistent with the stock line it's sold
on.  Now that regular stock lines can be set to 'any' department and the
department is just a filter for stock being put on sale, changing the
department of a stock type is allowed to enable mistakes to be fixed.

Closes #176 on github.