Discounts feature
Remove the interim discounts feature and the "convert transaction to
free drinks" feature.  They are both superceded by the new discounts
feature.

The amount of discount applied to an item is stored along with its
current price, allowing the original price to be restored and a new
discount computed when the discount policy is changed.

A description of the discount is stored per transaction line to enable
reporting by discount type.