When splitting a transaction, drag along voided lines
Transactions are split by the user marking a set of lines (that cannot
include payments or voided lines) and invoking the "split transaction"
function from the Manage Transaction menu.

Now that voided transaction lines can change value in open transactions
when the discount policy is changed, it's important that we never end up
in a state where a line that voids another line in the same transaction
moves to a separate transaction which is then closed first.

This commit ensures that chains of voids in the current transaction move
along to the split transaction along with the end of the chain.

This should be the final step in implementing discounts, and closes #151
on github.