New permissions / group system
All known permissions are added to the database, along with default
group membership.

Groups are now defined solely in the database, and not in the site
config file.  Groups should be removed from the site config after
migration.

There is a temporary new subcommand, "migrate-permissions", that
converts existing groups and permission grants to the new system.

The web interface has been updated to display new group membership.  In
the future, the web interface can make decisions based on users'
permissions since it now has the same group membership information as
the till.