Major units update
Implements #140 on github, including comments.

The Units table is expanded to include information about sale units, eg.
568ml being a "pint" and 750ml being a "wine bottle".  This information
is removed from the StockTypes table.

The reference to to the StockUnits table from StockItems is removed; the
stockunits table is now just a list of defaults to fill in the new
"description" and "size" columns in the stockitems table.

Modifiers will need updating after this change.