Delete annotations when stock item is deleted
If a delivery is manually changed from checked=true to checked=false,
it's possible for the user to delete a stock item that may already have
annotations, which will cause a crash.  Delete the annotations too.

This partially implements #166 on github.  In the future the
implementation can be changed to perform the delete cascade in the
database; not worth doing now because it would mean a schema change.