Set synchronize_session=False on update to stocktype stocktakes
sqlalchemy was trying to perform the 'ilike' function from the filter
locally to work out what is affected by the update.  This isn't
implemented, so crashes.  Set synchronize_session=False on the update so
it doesn't try, and explicitly leaves the session out of sync; we're
committing and flushing immediately afterwards so this doesn't matter.