Delay running lockscreen plugins by 2s
Plugins that take some time to run (eg. pinging a kitchen printer has a
1s blocking timeout; contacting the orders server may take a good part
of a second) are annoying if you are intending to move on to another
activity as soon as the lock screen starts.

Delay running the plugins by 2s when the lock screen starts up to give
attentive users a chance to move to another page without delay.