Use the middle section of the header for user messages
We used to use the middle of the header line to show a summary of what
was happening on pages other than the one currently displayed.  This
facility is currently unused, since register pages are destroyed when
not displayed.

This commit removes the vestigial "pagesummary" method and adds a ui
function to set the text in the middle of the header line.  This
function can be used by plugins, etc.