Better handling of deferral of part-paid transactions
Part-paid transactions now have the non-deferrable payments split out
into a separate transaction that stays in the current session when they
are deferred.  The payments transaction is balanced by a refund using
the default payment method (usually Cash) and a note is printed to wrap
the set-aside money.  Closes #204 on github.

Payment methods can now be marked as deferrable.  (There aren't any in
the base code, although online payments from the online ordering plugin
are now marked as deferrable.)  They stay with a transaction when it is
deferred.  Closes #212 on github.