Look up barcode details when processing scan
Look up barcode details in the barcodes table as soon as the scan is
received, and set the 'binding' attribute on the barcode object that's
passed to the input handler.

Other modules can now spot barcode objects by checking for the 'code'
attribute (similar to checking for the 'line' attribute on line keys)
and make use of the barcode attribute if set, without importing the
barcode module.