Change StockType.remaining to exclude stock on stocklines
StockType.remaining was only ever used to work out how much stock was
available on a continuous stock line.

Stock already attached to a regular or display stock line cannot be sold
through a continuous stock line. So it makes sense to exclude this when
calculating how much is available.

The same rule applies to stock sold directly through a stocktype (eg.
through a Barcode binding).