Don't conflate giving change with noninteractive payments
The Cash payment driver had a method "add_change()" which was used by
the Card driver for cashback, and by the register for the "Drink 'In'"
function and for deferring part-paid transactions. The presence of this
method was indicated by "change_given" being set to True.

This was confusing and was giving "change_given" two different meanings:
accepting overpayments, and supporting the "add_change()" method.

add_change() has been renamed to add_payment(), documented as being
noninteractive, and a new flag "add_payment_supported" has been added.
It's possible for this flag to be set on a payment method that doesn't
support overpayments.