Make opening the payment method menu on void optional
Add a config option to enable the automatic display of the payment
methods menu on voiding lines from a closed transaction optional.

This was originally introduced to stop people just hitting Enter and
getting a cash refund by mistake, but now that the default payment can
be configured to "None" it's not really needed.