Fix the buying list report
The meaning of StockType.instock had been changed to be more useful for
continuous stock lines, without realising that it was also used for the
buying list report.

Add StockType.all_instock with the previous behaviour, and change the
buying list function to use it.

Improve the buying list report to make use of Unit.stock_unit_name etc.

Closes #252 on github.