Enable in-progress Square transactions to be backgrounded
Sometimes Square takes a very long time to deal with cancelling a
transaction. If a user presses Recall Trans while a Square transaction
is in progress, we now hide the transaction and allow it to proceed in
the background.

This isn't a complete solution: the transaction will still need to be
resolved before the session can be ended, but it enables users to get on
with serving other customers instead of trapping them on the in-progress
transaction display.