Database notifications on stock changes
Adds triggers to send database notifications for changes to stock lines,
stock types and stock items.

Adds a command ('monitor') to output received notifications to the
console. This also listens for previously defined notifications.

This implements #276 on github. Apart from the 'monitor' command, the
first consumer of these notifications is expected to be the EMF 2024 bar
website.