Sort Square terminals by pairing time
When retrieving the list of terminals at a location, sort the list by
the terminal pairing time — this will feed through into the order the
terminals are displayed in menus.

(The Square API used to return terminals in a predictable order, even
though it was not documented to do so. Recently it started returning
them in random order in practice.)

Closes #280 on github.