Don't lock the register if a transaction closes within 3 seconds
If a Square transaction was in progress when the user locked the
register, it is likely to complete very quickly the next time the user
taps in. If this happens within 3 seconds of the user tapping in, don't
lock automatically — the user probably wants to start a new transaction.

See #288 on github.