Shortcut when Cash/Enter is pressed on an open zero-balance transaction
When Cash/Enter is pressed on an open zero-balance transaction, it
currently uses the default payment method if present, or pops up the
payment methods menu.

Popping up the menu is redundant — whatever choice the user makes, the
transaction will be closed in register.page.paymentkey(). Instead, we
just close the transaction immediately.

(The code to do this must also remain in paymentkey() because that
method can be invoked directly, or through explicity use of the payment
methods menu.)