Fixes to the new user passwords feature
Use "login" consistently, rather than a mix of "logon" and "login"

Use a mixin class for all login-related popups, so they are dismissed
when a new login is attempted

Fix the logic regarding forcing the user to set a password on login

Fix popups so that their layouts work in ncurses mode as well as gtk mode

Rework the login code path to avoid code duplication between register
and stockterminal — this also removes the need to pass a callback around
all of the login-related code.

Remove the "un-dismissable" feature of password setting popups.

Fix the "hidden" option on ui.editfield to work correctly when the user
moves the cursor position to the left

Revert adding user ID to the list of users, because it looks weird for
them not to be in user ID order! (They are in alphabetical order by full
user name.)