More password-related fixes
Ensure we don't pop up a password prompt or initial password setting
prompt on top of another user's page — our popups could be affected if
that page decided to deselect itself (eg. on timeout, or the user
logging in elsewhere).

If the user is logging in on top of their own page (eg. to unlock a
locked register page) that is fine. If they are logging in on top of
another user's page, we switch to the default page first.

Update the password setting popup to have a confirmation field for the
new password. Log password-change events where possible. (It isn't
possible to log a user's forced initial password setup, because at that
point there is no relevant active user to use for user.log())