Hide a non-fatal sqlalchemy log message
When postgresql is restarted, sqlalchemy emits a log message with level
ERROR while it cleans up the connection being used by the "listen"
module. This appears both as a toast and on stderr. It's harmless, but
it causes problems on the diskless tills — if the till produces any
output on stdout or stderr, they display it in a popup after the till
exits and wait for the user to confirm they have seen it.