Be more explicit about how we handle "None" in configuration settings
Prior to this change, a configuration setting could only ever be "None"
if that was the default, and an invalid value was present in the
database.

Now, only when support for None is explicitly requested, the default can
be anything (including None if wanted) and None will be returned when
there is an invalid value in the database.