Add a new configuration system for startup
Add a new configuration file that enables command-line argument defaults
to be set, a list of initial module imports to be specified, and a pool
of named database connections to be provided.

This is intended to replace the current configuration files in
/etc/quicktill/ although those are still supported for now.

If a new configuration file is present, files in /etc/quicktill/ will not
be read.

Closes #322 on github.