Improved stock take user interface
The stock take index page is now much faster to load. Closes #317 on
github.

Changes made on the "in progress" stock take page are now committed to the
database immediately, not just when the "Update" button is pressed.
Quantity adjustments are moved to a detail page, which frees up space and
should be more usable on mobile.

"Best-Before" dates can now be updated through the stock take interface.
Closes #328 on github.

The stock take interface now includes a field for "notes" on items of
stock. Closes #304 on github.

Attempting to add a stock item to a stock take that already contains that
stock item no longer crashes. Closes #306 on github.

The space freed up by the redesign closes #219 on github.

Everything mentioned in #203 is now implemented, so this closes #203 on
github.