Support marking transaction lines as ineligible for deletion
Some plugins create transaction lines at the same time as causing a side
effect, for example printing an order on a printer. Enable these lines to
be marked as "protected" so that they can only be voided, not deleted,
even if `register:max_transline_modify_age` time has not yet passed.

Closes #334 on github.